Fin refactor fichier gant.rs
